wip ab tests

+++ tests/ab/downloader.js
@@ -1,3 +1,4 @@
+const is_script = !module.parent;
 const Helper = require('../lib/helper');
 const path = require('path');
 const http = require("http");
@@ -7,9 +8,11 @@ const md5file = require('md5-file');
 const md5_hash = '0045f4c2a16ba5651f6e26139805d5b2';
 const assert = require('assert').strict;
 
-class Decipher {
-    constructor(dest) {
+class Downloader {
+    constructor(dest, port) {
+        mkdirp.sync(dest);
         this.dest = dest;
+        this.port = port;
         this.folder = Helper.getUploadServerFolder();
     }
 
@@ -24,7 +27,7 @@ class Decipher {
                         //console.log('download...');
                         fs.unlink(file, () => {
                             i -= 1;
-                            if(i <= 0) return resolve();
+                            if (i <= 0) return resolve();
                         })
                     });
                 });
@@ -37,35 +40,34 @@ class Decipher {
     }
 
     download(file, cb) {
-        http.get('http://localhost:3000/file/' + path.basename(file), res => {
+        http.get(`http://localhost:${this.port}/file/${path.basename(file)}`, res => {
             res.pipe(fs.createWriteStream(this.dest + '/' + path.basename(file))).on('finish', cb)
         })
     }
 };
 
-if (process.argv.includes('-d')) {
-    const destPos = process.argv.indexOf('-d') + 1;
-    const dest = process.argv[destPos];
-    mkdirp.sync(dest);
-    const decipher = new Decipher(dest);
+if (is_script) {
+    if (process.argv.includes('-d')) {
+        const dest = process.argv[process.argv.indexOf('-d') + 1];
+        const decipher = new Downloader(dest, 3000);
 
-    decipher.dec()
-        .then(() => {
+        decipher.dec()
+            .then(() => {
                 Helper.readDir(dest, (e, files) => {
                     console.log('Downloaded ' + files.length + ' files... now testing md5');
                     let i = files.length;
                     files.forEach((file) => {
                         md5file(file, (e, md5) => {
-                            if(e) throw e;
+                            if (e) throw e;
                             assert.deepStrictEqual(md5, md5_hash);
                             fs.unlinkSync(file);
                             i -= 1;
-                            if(i <= 0 ) console.log('test finished');
+                            if (i <= 0) console.log('test finished');
                         })
                     });
-            })
-        });
-
+                })
+            });
+    }
 } else {
-    module.exports = Decipher;
+    module.exports = Downloader;
 }
